ui/src/requirements.txt
Flask==3.1.0
GitPython==3.1.43
GitPython==3.1.43
PyYAML==6.0.2
PyYAML==6.0.2
